Working up test harness code for the async daemon
