+++ src/Marten.Testing/Events/Daemon/TestingSupport/Arrival.cs
@@ -1,0 +1,8 @@
+namespace Marten.Testing.Events.Daemon.TestingSupport
+{
+    public class Arrival
+    {
+        public int Day { get; set; }
+        public string State { get; set; }
+    }
+}
