+++ src/Marten.Testing/Events/Daemon/TestingSupport/Departure.cs
@@ -1,0 +1,8 @@
+namespace Marten.Testing.Events.Daemon.TestingSupport
+{
+    public class Departure
+    {
+        public int Day { get; set; }
+        public string State { get; set; }
+    }
+}
