+++ src/Marten.Testing/Events/Daemon/TestingSupport/Direction.cs
@@ -1,0 +1,10 @@
+namespace Marten.Testing.Events.Daemon.TestingSupport
+{
+    public enum Direction
+    {
+        North,
+        South,
+        East,
+        West
+    }
+}
