+++ src/Marten.Testing/Events/Daemon/TestingSupport/Movement.cs
@@ -1,0 +1,14 @@
+namespace Marten.Testing.Events.Daemon.TestingSupport
+{
+    public class Movement
+    {
+        public Direction Direction
+        {
+            get;
+            set;
+
+        }
+
+        public double Distance { get; set; }
+    }
+}
