+++ src/Marten.Testing/Events/Daemon/TestingSupport/TripAborted.cs
@@ -1,0 +1,7 @@
+namespace Marten.Testing.Events.Daemon.TestingSupport
+{
+    public class TripAborted
+    {
+
+    }
+}
