+++ src/Marten.Testing/Events/Daemon/TestingSupport/TripEnded.cs
@@ -1,0 +1,8 @@
+namespace Marten.Testing.Events.Daemon.TestingSupport
+{
+    public class TripEnded
+    {
+        public int Day { get; set; }
+        public string State { get; set; }
+    }
+}
