+++ src/Marten.Testing/Events/Daemon/TestingSupport/TripStarted.cs
@@ -1,0 +1,7 @@
+namespace Marten.Testing.Events.Daemon.TestingSupport
+{
+    public class TripStarted
+    {
+        public int Day { get; set; }
+    }
+}
