+++ src/Marten/Events/Daemon/Progress/UpdateProjectionProgress.cs
@@ -35,7 +35,7 @@ namespace Marten.Events.Daemon.Progress
             parameters[2].NpgsqlDbType = NpgsqlDbType.Bigint;
         }
 
-        public Type DocumentType => typeof(ProjectionProgress);
+        public Type DocumentType => typeof(IEvent);
         public void Postprocess(DbDataReader reader, IList<Exception> exceptions)
         {
             if (reader.RecordsAffected != 1)
