+++ src/Marten/Events/Operations/EstablishTombstoneStream.cs
@@ -65,7 +65,7 @@ DO NOTHING
             _configureParameter(parameters[0]);
         }
 
-        public Type DocumentType => typeof(StreamAction);
+        public Type DocumentType => typeof(IEvent);
         public void Postprocess(DbDataReader reader, IList<Exception> exceptions)
         {
             // Nothing
