+++ src/Marten/Events/Operations/EventProgressWrite.cs
@@ -31,7 +31,7 @@ namespace Marten.Events.Operations
             builder.Append($"select {_sproc}(:{nameArg.ParameterName}, :{numberArg.ParameterName})");
         }
 
-        public Type DocumentType => null;
+        public Type DocumentType => typeof(IEvent);
 
         public void Postprocess(DbDataReader reader, IList<Exception> exceptions)
         {
