+++ src/Marten/Events/Operations/InsertStreamBase.cs
@@ -22,7 +22,7 @@ namespace Marten.Events.Operations
 
         public abstract void ConfigureCommand(CommandBuilder builder, IMartenSession session);
 
-        public Type DocumentType => typeof(StreamAction);
+        public Type DocumentType => typeof(IEvent);
         public void Postprocess(DbDataReader reader, IList<Exception> exceptions)
         {
         }
